Change checked Disk from LogicalDisk to Volume
The reason behind the change is that it seems that `LogicalVolume` will partly consider disk quotas and partly ignore them. We had an issue where a customer reported negative values for the calculated free space in case they had quotas for the checking user in place. Looks like `Size` of a `LogicalDisk` gives the size of the quota and `FreeSpace` ignores quota.
So I changed the queried classes to hope for total values without queries.

@LordHepipud : Please feel free to correct me if I went the wrong path. And: I can think of reasons why you used `LogicalDisk` instead of `Volume`, so if there's anything I missed, please feel free to just make "User running the plugin must not have a quota applied" a requirement and throw away my PR.

Some things to note about this PR:

* @RincewindsHat had the initial idea about what to look for
* I did something that I would *never* advise someone to do but I checked *first* with the maintainer if it's ok as an exception (details follow)
* I don't know how to code PowerShell
* I used an LLM to help me with identifying the new code snippets I needed to use
* I don't have a way to check the code because I don't use Windows